refactor: extract gateway admin audit domain
